LimitedPriorityKdTree renamed to SegmentationTree Attributes data type changed from int to double Concept of BattleTime replaced by roundNum + roundTime enemy predicate path drawing improved Movement classifier concept introduced druss gun tweaks

+++ tomcat/src/lxx/model/attributes/attribute_extractors/AttributeValueExtractor.java
@@ -15,6 +15,6 @@ import java.util.List;
  */
 public interface AttributeValueExtractor {
 
-    public int getAttributeValue(LXXRobot enemy, LXXRobot me, List<LXXBullet> myBullets);
+    public double getAttributeValue(LXXRobot enemy, LXXRobot me, List<LXXBullet> myBullets);
 
 }
